requirements.txt
wheel
setuptools
pytesseract
pillow
jamdict
git+https://github.com/neocl/jamdict_data.git@refs/pull/2/merge
pyautogui
fire
hsluv
colorama
